Move the User namespace into a new Models namespace
