+++ tests/Feature/StaticPagesTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature;
 
-use App\User;
+use App\Models\User;
 use Tests\TestCase;
 
 class StaticPagesTest extends TestCase
